* added: HiResSampler und DirectScheduler * prepared: batterysim * changed: 1 month pre-sim * refactoring

+++ run_pre.py
@@ -20,12 +20,17 @@ if os.path.exists(sim_dfn):
 sam_dfn = str(os.path.join(d, '.'.join((str(sc.seed), 'samples', 'npy'))))
 if os.path.exists(sam_dfn):
     raise RuntimeError('File already exists: %s' % sam_dfn)
+mod_dfn = str(os.path.join(d, '.'.join((str(sc.seed), 'modes', 'npy'))))
+if os.path.exists(mod_dfn):
+    raise RuntimeError('File already exists: %s' % mod_dfn)
 
-sim_data, sample_data = simulator.run_pre(sc)
+sim_data, modes_data, sample_data = simulator.run_pre(sc)
 np.save(sim_dfn, sim_data)
 np.save(sam_dfn, sample_data)
+np.save(mod_dfn, modes_data)
 sc.run_pre_datafile = sim_dfn
 sc.run_pre_samplesfile = sam_dfn
+sc.run_pre_modesfile = mod_dfn
 sc.save_JSON(sc_file)
 
 # import matplotlib.pyplot as plt
